@@ -44,100 +51,126 @@
         "react-intl": "^5.8.8",
         "react-maskinput": "^1.0.1",
         "react-perfect-scrollbar": "^1.5.8",
-        "react-redux": "7.1.0",
+        "react-redux": "^7.2.1",
         "react-router": "^5.1.2",
         "react-router-dom": "^5.1.2",
-        "react-scripts": "3.4.3",
+        "react-string-format": "^0.0.6",
         "recharts": "^1.8.5",
-        "redux": "^4.0.1",
+        "redux": "^4.0.5",
         "redux-saga": "^1.1.3",
         "redux-sentry-middleware": "^0.1.8",
+        "sass-loader": "^9.0.3",
         "sinon": "^7.2.3",
+        "style-loader": "^1.2.1",
         "stylelint": "^13.2.1",
         "stylelint-config-standard": "^20.0.0",
+        "stylelint-no-unsupported-browser-features": "^4.0.0",
         "stylelint-order": "^4.0.0",
-        "tslint": "^5.13.1",
+        "stylelint-prettier": "^1.1.2",
+        "stylelint-scss": "^3.18.0",
+        "stylelint-selector-bem-pattern": "^2.1.0",
+        "thread-loader": "^2.1.3",
+        "ts-loader": "^8.0.2",
         "ua-parser-js": "^0.7.19",
+        "url-loader": "^4.1.1",
         "ws": "^7.2.0"
     },
     "scripts": {
-        "build:css": "postcss src/index.pcss -o src/index.css",
-        "watch:css": "postcss src/index.pcss -o src/index.css -w",
-        "lint": "yarn style:lint && tsc --noEmit && tslint --project .",
-        "lint:fix": "tslint --project . --fix",
+        "lint": "eslint --ext .ts,.tsx src",
+        "lint:fix": "eslint --ext .ts,.tsx src --fix",
         "style:lint": "npx stylelint '**/*.pcss'",
         "mockAPI": "node mocks/mockAPI.js",
         "mockWS": "node mocks/mockWS.js",
-        "start": "npm-run-all -p watch:css start-js",
-        "start-mock": "npm-run-all -p watch:css start-js mockAPI mockWS",
-        "start-js": "react-scripts start",
-        "build-js": "react-app-rewired build --scripts-version react-scripts build",
-        "build": "npm-run-all build:css build-js clean:mapfiles",
-        "clean:mapfiles": "find build -name '*.map' -exec rm {} ';'",
-        "test": "react-scripts test",
-        "test:coverage": "react-scripts test --collectCoverage=true",
-        "test:ci": "react-scripts test --ci",
+        "start": "cross-env NODE_OPTIONS=--max_old_space_size=2048 webpack-dev-server --config ./webpack/development.ts",
+        "start-mock": "npm-run-all -p start mockAPI mockWS",
+        "build": "cross-env NODE_ENV=production && cross-env NODE_OPTIONS=--max_old_space_size=4096 webpack --optimize-minimize --config ./webpack/production.ts",
+        "test": "jest --config jest.config.js --silent",
+        "test:coverage": "jest --config jest.config.js --collectCoverage=true",
+        "test:ci": "jest --config jest.config.js --silent --ci",
         "spec:integration": "cypress open",
-        "eject": "react-scripts eject",
-        "clean": "rimraf build",
-        "analyze": "source-map-explorer build/static/js/main.*"
+        "clean": "rimraf build"
     },
     "devDependencies": {
         "@openware/coding-standards": "2.3.0",
         "@types/axios": "^0.14.0",
         "@types/classnames": "^2.2.6",
+        "@types/copy-webpack-plugin": "^6.0.0",
         "@types/enzyme": "^3.1.13",
         "@types/enzyme-adapter-react-16": "^1.0.3",
         "@types/history": "^4.7.0",
-        "@types/jest": "^23.3.1",
+        "@types/jest": "^26.0.10",
         "@types/jquery": "^3.3.6",
+        "@types/mini-css-extract-plugin": "^0.9.1",
         "@types/node": "^10.9.4",
+        "@types/qrcode.react": "^1.0.1",
         "@types/react": "^16.9.14",
         "@types/react-dom": "^16.9.4",
         "@types/react-hot-loader": "^4.1.0",
-        "@types/react-redux": "7.1.9",
+        "@types/react-redux": "^7.1.9",
         "@types/react-router": "^5.1.3",
         "@types/react-router-dom": "^5.1.3",
         "@types/react-test-renderer": "^16.0.2",
-        "@types/webpack-env": "^1.13.6",
-        "autoprefixer": "^9.1.1",
+        "@types/redux-form": "^8.2.7",
+        "@types/webpack": "^4.41.21",
+        "@types/webpack-dev-server": "^3.11.0",
+        "@types/webpack-env": "^1.15.2",
+        "@types/webpack-merge": "^4.1.5",
+        "autoprefixer": "^10.0.1",
         "axios-mock-adapter": "^1.16.0",
-        "babel-jest": "24.9.0",
         "chalk": "^2.4.1",
         "classnames": "^2.2.6",
         "colors": "1.3.2",
+        "copy-webpack-plugin": "^6.2.1",
+        "cross-env": "^7.0.2",
         "cryptocurrency-icons": "^0.16.1",
         "cssnano": "^4.0.5",
         "cypress": "^3.1.5",
         "enzyme": "^3.10.0",
         "enzyme-adapter-react-16": "^1.10.0",
         "enzyme-to-json": "^3.3.4",
+        "eslint": "^7.7.0",
+        "eslint-import-resolver-typescript": "^2.2.1",
+        "eslint-import-resolver-webpack": "^0.12.2",
+        "eslint-plugin-import": "^2.22.0",
         "file-loader": "^1.1.11",
+        "fork-ts-checker-webpack-plugin": "^5.1.0",
         "header-case-normalizer": "1.0.3",
+        "html-webpack-plugin": "^4.3.0",
         "husky": "^1.0.0-rc.13",
         "isomorphic-ws": "^4.0.1",
-        "jest": "24.9.0",
+        "jest": "^26.4.1",
+        "jest-config": "^26.4.1",
+        "jest-environment-enzyme": "^7.1.2",
+        "jest-enzyme": "^7.1.2",
         "js-combinatorics": "0.5.0",
+        "mini-css-extract-plugin": "^0.10.0",
         "npm-run-all": "^4.1.5",
+        "optimize-css-assets-webpack-plugin": "^5.0.3",
+        "postcss": "^8.1.4",
         "postcss-autoreset": "^2.0.2",
-        "postcss-hexrgba": "^1.0.1",
-        "postcss-import": "^12.0.0",
-        "postcss-nested": "^3.0.0",
-        "prettier": "1.9.0",
+        "postcss-custom-properties": "^10.0.0",
+        "postcss-hexrgba": "^2.0.1",
+        "postcss-import": "^13.0.0",
+        "postcss-loader": "^4.0.4",
+        "postcss-nested": "^5.0.1",
         "ramda": "^0.26.1",
-        "react-app-rewired": "^2.1.5",
         "react-hot-loader": "^4.3.4",
         "react-test-renderer": "^16.4.2",
         "redux-mock-store": "^1.5.3",
         "rimraf": "^2.6.2",
         "source-map-explorer": "^1.7.0",
-        "ts-jest": "^23.1.3",
-        "ts-node": "^7.0.1",
-        "tslint-plugin-prettier": "^2.3.0",
-        "tslint-react-hooks": "^2.2.2",
-        "typescript": "3.8.2",
-        "webpack": "4.42.0",
-        "webpack-obfuscator": "^2.3.0"
+        "speed-measure-webpack-plugin": "^1.3.3",
+        "stylelint-config-prettier": "^8.0.2",
+        "stylelint-config-recommended": "^3.0.0",
+        "ts-jest": "^26.2.0",
+        "ts-node": "^9.0.0",
+        "typescript": "^3.9.7",
+        "webpack": "^4.44.1",
+        "webpack-bundle-analyzer": "^3.8.0",
+        "webpack-cli": "^3.3.12",
+        "webpack-dev-server": "^3.11.0",
+        "webpack-merge": "^5.1.2",
+        "webpack-obfuscator": "^0.18.0"
     },
     "browserslist": [
         ">0.2%",
