@@ -1,4 +1,5 @@
 import { mount, ReactWrapper, shallow, ShallowWrapper } from 'enzyme';
+import { TestComponentWrapper } from 'lib/test';
 import * as React from 'react';
 import { spy } from 'sinon';
 import { Markets, MarketsProps } from '.';
@@ -16,32 +17,42 @@ const defaultProps: MarketsProps = {
 };
 
 const setup = (props?: Partial<MarketsProps>) =>
-    shallow(<Markets {...{ ...defaultProps, ...props }} />);
+    shallow(
+        <TestComponentWrapper>
+            <Markets {...{ ...defaultProps, ...props }} />
+        </TestComponentWrapper>
+    );
 
-describe.skip('Markets', () => {
-   let wrapper: ShallowWrapper;
+describe('Markets', () => {
+    let wrapper: ShallowWrapper;
 
-   beforeEach(() => {
-      wrapper = setup();
-   });
+    beforeEach(() => {
+        wrapper = setup();
+    });
 
-   it('should render', () => {
-       expect(wrapper).toMatchSnapshot();
-   });
+    it('should render', () => {
+        expect(wrapper.render()).toMatchSnapshot();
+    });
 
-   it('should render empty data', () => {
-       wrapper = setup({ data: [] });
-       expect(wrapper).toMatchSnapshot();
-   });
+    it('should render empty data', () => {
+        wrapper = setup({ data: [] });
+        expect(wrapper).toMatchSnapshot();
+    });
 
-   it('should set selected market in props', () => {
-      const keyIndex = 0;
-      const selectedKey = 'ETH/LTC';
-      const component: ReactWrapper = mount(
-         <Markets data={data} rowKeyIndex={keyIndex} selectedKey={selectedKey} onSelect={onSelect} />,
-      );
-      const resultSelectedRow = '<tr class="cr-table__row--selected"><td>ETH/LTC</td><td>0.223100</td><td><span class="__positive">+25.00%</span></td></tr>';
-      const expectedSelectedRow = component.find('.cr-table__row--selected').first().html();
-      expect(resultSelectedRow).toBe(expectedSelectedRow);
-   });
+    it('should set selected market in props', () => {
+        const keyIndex = 0;
+        const selectedKey = 'ETH/LTC';
+
+        const component = setup({
+            data,
+            rowKeyIndex: keyIndex,
+            selectedKey,
+            onSelect,
+        }).render();
+
+        const resultSelectedRow = '<td>ETH/LTC</td><td>0.223100</td><td><span class="__positive">+25.00%</span></td>';
+        const expectedSelectedRow = component.find('.cr-table__row--selected').first().html();
+        console.log(expectedSelectedRow);
+        expect(resultSelectedRow).toBe(expectedSelectedRow);
+    });
 });
