@@ -1,8 +1,9 @@
 import classnames from 'classnames';
 import * as React from 'react';
-import { CellData, FilterInput, Table } from '../';
 import { DEFAULT_MARKET_HEADERS } from '../../constants';
 import { hasDuplicates } from '../../helpers';
+import { FilterInput } from '../FilterInput';
+import { CellData, Table } from '../Table';
 
 export interface MarketsProps {
     /**
@@ -45,11 +46,14 @@ export const Markets = (props: MarketsProps) => {
 
     const { headers, title, filterPlaceholder = '', rowKeyIndex, selectedKey, data, filters } = props;
 
-    const searchFilter = React.useCallback((row: CellData[], key: string) => {
-        setSearchKey(key);
+    const searchFilter = React.useCallback(
+        (row: CellData[], key: string) => {
+            setSearchKey(key);
 
-        return (row[0] as string).toLowerCase().includes(searchKey.toLowerCase());
-    }, [searchKey]);
+            return (row[0] as string).toLowerCase().includes(searchKey.toLowerCase());
+        },
+        [searchKey]
+    );
 
     const renderChange = React.useCallback((cell: string) => {
         const isItChangeValue = (c: string) => {
@@ -64,51 +68,63 @@ export const Markets = (props: MarketsProps) => {
         return <span className={className}>{cell}</span>;
     }, []);
 
-    const mapRows = React.useCallback((cell: CellData) => {
-        const isChangeValue = typeof(cell) === 'string' && (cell.charAt(0) === '-' || cell.charAt(0) === '+');
+    const mapRows = React.useCallback(
+        (cell: CellData) => {
+            const isChangeValue = typeof cell === 'string' && (cell.charAt(0) === '-' || cell.charAt(0) === '+');
 
-        return isChangeValue ? renderChange(cell as string) : cell;
-    }, [renderChange]);
+            return isChangeValue ? renderChange(cell as string) : cell;
+        },
+        [renderChange]
+    );
 
-    const filterType = React.useCallback((headerKey: string, key: string) => (item: CellData[]) => {
-        const typeIndex = (props.headers || DEFAULT_MARKET_HEADERS).indexOf(headerKey);
+    const filterType = React.useCallback(
+        (headerKey: string, key: string) => (item: CellData[]) => {
+            const typeIndex = (props.headers || DEFAULT_MARKET_HEADERS).indexOf(headerKey);
 
-        return (item[typeIndex] as string).includes(key);
-    }, [props.headers]);
+            return (item[typeIndex] as string).includes(key);
+        },
+        [props.headers]
+    );
 
     const createUniqueCurrencies = React.useCallback((currencies: string[], market: string) => {
         const marketCurrencies = market.split('/').map((c: string) => c.trim());
-        const uniqueCurrencies = marketCurrencies.filter(c => !hasDuplicates(currencies, c));
+        const uniqueCurrencies = marketCurrencies.filter((c) => !hasDuplicates(currencies, c));
 
         return currencies.concat(uniqueCurrencies);
     }, []);
 
-    const transformCurrencyToFilter = React.useCallback((currency: string) => ({
-        name: currency,
-        filter: filterType('Pair', currency),
-    }), [filterType]);
+    const transformCurrencyToFilter = React.useCallback(
+        (currency: string) => ({
+            name: currency,
+            filter: filterType('Pair', currency),
+        }),
+        [filterType]
+    );
 
     const getFilters = React.useCallback(() => {
-        const currencyFilters = data && data.length > 0
-            ? data
-                .map((market: React.ReactNode[]) => market[0] as string)
-                .reduce(createUniqueCurrencies, [])
-                .map(transformCurrencyToFilter)
-            : [];
+        const currencyFilters =
+            data && data.length > 0
+                ? data
+                      .map((market: React.ReactNode[]) => market[0] as string)
+                      .reduce(createUniqueCurrencies, [])
+                      .map(transformCurrencyToFilter)
+                : [];
 
-        return filters ? [
-            {
-                name: 'All',
-                filter: filterType('Pair', ''),
-            },
-            ...currencyFilters,
-        ] : [];
+        return filters
+            ? [
+                  {
+                      name: 'All',
+                      filter: filterType('Pair', ''),
+                  },
+                  ...currencyFilters,
+              ]
+            : [];
     }, [createUniqueCurrencies, filterType, transformCurrencyToFilter, data, filters]);
 
     const getTableData = React.useMemo(() => {
         const fd = data.filter(w => (w[0] as string).toLowerCase().includes(searchKey.toLowerCase()));
 
-        return fd.map(row => row.map(mapRows));
+        return fd.map((row) => row.map(mapRows));
     }, [data, mapRows, searchKey]);
 
     return (
@@ -122,11 +138,7 @@ export const Markets = (props: MarketsProps) => {
                 onSelect={props.onSelect}
                 titleComponent={title || 'Markets'}
             />
-            <FilterInput
-                data={props.data}
-                filter={searchFilter}
-                placeholder={filterPlaceholder}
-            />
+            <FilterInput data={props.data} filter={searchFilter} placeholder={filterPlaceholder} />
         </div>
     );
 };
