@@ -1,8 +1,8 @@
 //tslint:disable-next-line
-import * as QRCodeGenerator from 'qrcode.react';
-import * as React from 'react';
+import QRCodeGenerator from 'qrcode.react';
+import React from 'react';
 
-interface QRCodeProps {
+export interface QRCodeProps {
     /**
      * Data which is used to generate QR code(e.g. wallet address).
      * @default Required
@@ -18,14 +18,10 @@ interface QRCodeProps {
 /**
  * Component for displaying QR code.
  */
-const QRCodeComponent: React.FC<QRCodeProps> = ({ data = '', dimensions }) => {
+const QRCodeComponent: React.FC<QRCodeProps> = ({ data = '', dimensions = 0 }) => {
     return (
         <div className="qr-code">
-            <QRCodeGenerator
-                value={data}
-                size={dimensions}
-                renderAs="svg"
-            />
+            <QRCodeGenerator value={data} size={dimensions} renderAs="svg" />
         </div>
     );
 };
