@@ -1,5 +1,5 @@
 import cr from 'classnames';
-import * as moment from 'moment';
+import moment from 'moment';
 import * as React from 'react';
 import { Button } from 'react-bootstrap';
 import { injectIntl } from 'react-intl';
