@@ -1,4 +1,4 @@
-import * as moment from 'moment';
+import moment from 'moment';
 import { localeDate } from './localeDate';
 
 export const sortByDate = (prop: string, format: string) => {
