@@ -5,8 +5,7 @@ import {
   MapDispatchToPropsFunction,
   MapStateToProps,
 } from 'react-redux';
-import { RouterProps } from 'react-router';
-import { withRouter } from 'react-router-dom';
+import { RouterProps, withRouter } from 'react-router';
 import { compose } from 'redux';
 import { IntlProps } from '../../';
 import { ChangePassword } from '../../components';
