@@ -1,25 +1,20 @@
 import { shallow } from 'enzyme';
-import * as React from 'react';
-import { connect, Provider } from 'react-redux';
-import { createStore } from 'redux';
+import React from 'react';
+import { TestComponentWrapper } from 'lib/test';
 import { IntlProps } from '../../';
-import { rootReducer } from '../../modules';
 import { ChangeForgottenPasswordScreen } from '../ChangeForgottenPasswordScreen';
 
-const store = createStore(rootReducer);
-const ChangeForgottenPasswordTab = connect()(ChangeForgottenPasswordScreen);
 
 const setup = (props: Partial<IntlProps> = {}) =>
     shallow(
-        <Provider store={store}>
-            <ChangeForgottenPasswordTab/>
-        </Provider>,
+        <TestComponentWrapper>
+            <ChangeForgottenPasswordScreen />
+        </TestComponentWrapper>
     );
 
 describe('ChangeForgottenPasswordScreen test', () => {
     it('should render', () => {
-        const wrapper = setup();
+        const wrapper = setup().render();
         expect(wrapper).toMatchSnapshot();
-        expect(wrapper).toBeDefined();
     });
 });
