@@ -1,24 +1,19 @@
 import { shallow } from 'enzyme';
-import * as React from 'react';
-import { Provider } from 'react-redux';
-import { createStore } from 'redux';
+import React from 'react';
+import { TestComponentWrapper } from 'lib/test';
 import { IntlProps } from '../../';
-import { rootReducer } from '../../modules';
 import { RestrictedScreen } from '../RestrictedScreen';
-
-const store = createStore(rootReducer);
 
 const setup = (props: Partial<IntlProps> = {}) =>
     shallow(
-        <Provider store={store}>
+        <TestComponentWrapper>
             <RestrictedScreen />
-        </Provider>,
+        </TestComponentWrapper>
     );
 
 describe('RestrictedScreen', () => {
-    const wrapper = setup();
-
     it('should render', () => {
+        const wrapper = setup().render();
         expect(wrapper).toMatchSnapshot();
     });
 });
