@@ -1,25 +1,19 @@
 import { shallow } from 'enzyme';
-import * as React from 'react';
-import { connect, Provider } from 'react-redux';
-import { createStore } from 'redux';
+import React from 'react';
+import { TestComponentWrapper } from 'lib/test';
 import { IntlProps } from '../../';
-import { rootReducer } from '../../modules';
 import { SignInScreen } from '../SignInScreen';
-
-const store = createStore(rootReducer);
-const Identity = connect()(SignInScreen);
 
 const setup = (props: Partial<IntlProps> = {}) =>
     shallow(
-        <Provider store={store}>
-            <Identity />
-        </Provider>,
+        <TestComponentWrapper>
+            <SignInScreen />
+        </TestComponentWrapper>
     );
 
 describe('SignInScreen', () => {
-    const wrapper = setup();
-
     it('should render', () => {
+        const wrapper = setup().render();
         expect(wrapper).toMatchSnapshot();
     });
 });
